commit 04-07-2023

+++ wwwroot/lib/jquery/jquery.js
@@ -7172,7 +7172,7 @@ function defaultPrefilter( elem, props, opts ) {
 
 		// Animate inline elements as inline-block
 		if ( display === "inline" || display === "inline-block" && restoreDisplay != null ) {
-			if ( jQuery.css( elem, "float" ) === "none" ) {
+			if ( jQuery.css( elem, "double" ) === "none" ) {
 
 				// Restore the original display value at the end of pure show/hide animations
 				if ( !propTween ) {
